Add special handling for an "echo" command / "OTA".
The command information is sent back to IoTConnect in a test_last_command STRING attribute.
The OTA status is sent back to IoTConnect in a test_ota_status STRING attribute.
Note device templates have to be updated to include the:
- echo command (with parameter / acknowledgement)
- test_last_command STRING attribute
- test_ota_status STRING attribute
Even without the test code, the device will print to the console.